remove `evaluate-literal-unaryexpression.js` to improve codecov
