+++ rules/consistent-indexof-check.js
@@ -1,5 +1,4 @@
 'use strict';
-const evaluateLiteralUnaryExpression = require('./utils/evaluate-literal-unaryexpression.js');
 const {isMethodCall} = require('./ast/index.js');
 
 const MESSAGE_ID = 'consistent-indexof-check';
@@ -13,6 +12,41 @@ const comparisonMap = {
 	'>': '<',
 	'>=': '<=',
 };
+
+/**
+Evaluate the value of a UnaryExpression node which contain a Literal node.
+Make sure to call this function only when you are sure that the node is a UnaryExpression node or Literal node.
+
+@param {import('estree').Node} node
+@returns
+*/
+function evaluateLiteralUnaryExpression(node) {
+	// Base case: if the argument is a numeric literal, return its value directly
+	if (node.type === 'Literal') {
+		return node.value;
+	}
+
+	// If the argument is another UnaryExpression, recursively evaluate its value
+	if (node.type === 'UnaryExpression') {
+		const argumentValue = evaluateLiteralUnaryExpression(node.argument);
+
+		switch (node.operator) {
+			case '-': {
+				return -argumentValue;
+			}
+
+			case '+': {
+				return Number(argumentValue);
+			}
+
+			case '~': {
+				return ~argumentValue; // eslint-disable-line no-bitwise
+			}
+
+			default:
+		}
+	}
+}
 
 /**
 Determine the appropriate replacement based on the operator and value.
@@ -55,10 +89,6 @@ Find and process references to a given identifier.
 */
 function findAndProcessReferences(context, identifierNode) {
 	const variable = context.sourceCode.getDeclaredVariables(identifierNode.parent).find(variable => variable.name === identifierNode.name);
-
-	if (!variable) {
-		return;
-	}
 
 	for (const reference of variable.references) {
 		if (reference.identifier === identifierNode) {
